MenuBar and functionality
Created methods to close app, go back to main menu and back to game lobby and created menuBar with menuItems to use them. Not all methods are implemented everywhere, but they are implemented depending on what functionality is needed in the different classes. For example the main menu page should only have the functionality to close out of these new methods.